# backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CNote API Documentation",
        "version": "1.0.0",
        "description": "This is the API documentation for the CNote backend."
    },
    "servers": [
        {
            "url": "http://localhost:5000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "summary": "Register a new user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["username", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Username already exists"
                    },
                    "500": {
                        "description": "Failed to register user"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "summary": "Log in a user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": { "type": "string" },
                                    "password": { "type": "string" }
                                },
                                "required": ["username", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": { "type": "string" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "500": {
                        "description": "Failed to log in"
                    }
                }
            }
        },
        "/folders": {
            "get": {
                "summary": "Get all folders",
                "responses": {
                    "200": {
                        "description": "A list of folders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "_id": { "type": "string" },
                                            "name": { "type": "string" },
                                            "parentFolder": { "type": "string", "nullable": true },
                                            "notes": {
                                                "type": "array",
                                                "items": { "type": "string" }
                                            },
                                            "user": { "type": "string" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "summary": "Create a new folder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": { "type": "string" },
                                    "parentFolder": { "type": "string", "nullable": true }
                                },
                                "required": ["name"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Folder created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": { "type": "string" },
                                        "name": { "type": "string" },
                                        "parentFolder": { "type": "string", "nullable": true },
                                        "notes": {
                                            "type": "array",
                                            "items": { "type": "string" }
                                        },
                                        "user": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Failed to create folder"
                    }
                }
            }
        },
        "/folders/{id}": {
            "get": {
                "summary": "Get a folder by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "Folder ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": { "type": "string" },
                                        "name": { "type": "string" },
                                        "parentFolder": { "type": "string", "nullable": true },
                                        "notes": {
                                            "type": "array",
                                            "items": { "type": "string" }
                                        },
                                        "user": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Folder not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "summary": "Delete a folder by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "Folder ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Folder deleted successfully"
                    },
                    "404": {
                        "description": "Folder not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/notes": {
            "get": {
                "summary": "Get all notes",
                "responses": {
                    "200": {
                        "description": "A list of notes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "_id": { "type": "string" },
                                            "title": { "type": "string" },
                                            "keyPoints": { "type": "string" },
                                            "detailedNotes": { "type": "string" },
                                            "summary": { "type": "string" },
                                            "user": { "type": "string" }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "summary": "Create a new note",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": { "type": "string" },
                                    "keyPoints": { "type": "string" },
                                    "detailedNotes": { "type": "string" },
                                    "summary": { "type": "string" },
                                    "folderId": { "type": "string", "nullable": true }
                                },
                                "required": ["title", "detailedNotes"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": { "type": "string" },
                                        "title": { "type": "string" },
                                        "keyPoints": { "type": "string" },
                                        "detailedNotes": { "type": "string" },
                                        "summary": { "type": "string" },
                                        "user": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Failed to create note"
                    }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "summary": "Get a note by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "Note ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": { "type": "string" },
                                        "title": { "type": "string" },
                                        "keyPoints": { "type": "string" },
                                        "detailedNotes": { "type": "string" },
                                        "summary": { "type": "string" },
                                        "user": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "put": {
                "summary": "Update a note by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "Note ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": { "type": "string" },
                                    "keyPoints": { "type": "string" },
                                    "detailedNotes": { "type": "string" },
                                    "summary": { "type": "string" }
                                },
                                "required": ["title", "detailedNotes"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "_id": { "type": "string" },
                                        "title": { "type": "string" },
                                        "keyPoints": { "type": "string" },
                                        "detailedNotes": { "type": "string" },
                                        "summary": { "type": "string" },
                                        "user": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "summary": "Delete a note by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" },
                        "description": "Note ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note deleted successfully"
                    },
                    "404": {
                        "description": "Note not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    }
}